# Implement Data.Graph.transposeG with STArray