@treeowl 2023-01-27
I would have expected the simple version to optimize (by list fusion) into the fast one. I think it would be worth figuring out why it doesn't. Do we need more `INLINABLE` and/or `INLINE` pragmas somewhere?