@meooow25 2023-01-28
I didn't think of that 🤦 
It does work just as well after making things inline! I'll make another PR.