@treeowl 2023-01-28
My guess is that we want

```haskell
{-# INLINE reverseE #-}
{-# NOINLINE transposeG #-}
{-# INLINABLE buildG #-}
```

but it's also possible we want `{-# INLINE buildG #-}`. These pragmas are really hard for my poor brain.