@meooow25 2023-01-29
I also need to inline `edges` to the optimized result. Why do we want the `NOINLINE` on `transposeG` though?