@treeowl 2023-01-29
Two reasons:

1. I want to lock in the savings before prices increase. That is, I'd like to make sure the optimizer doesn't mess up what you've optimized in a more interesting context.
2. I don't see much potential for inlining it to help. The only way I could see would be if there were some special `RULES` for folding over an array constructed by `listArray`, and if someone passed such to `transposeG`. But there is no such rule in `GHC.Arr`, and I would be a bit surprised if it were common to build graphs using `listArray` anyway. So why should the optimizer even waste its time considering inlining `transposeG`?

These reasons aren't dispositive; they're just my thoughts on the matter. If you disagree, feel free to leave that pragma out.